pi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyserial==3.5
picamera2==0.3.21
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
ultralytics==8.3.0
face-recognition==1.3.0
numpy==1.26.4
Pillow==10.4.0
pydantic==2.8.2
aiofiles==24.1.0
anthropic==0.40.0
